Remove exodvs and deploy mainnet (#7)
* Update subquery-node image version in docker-compose

The docker-compose file has been updated to use a newer version of the subquery-node image. The version has been updated from v3.2.0 to v3.11.2 to leverage new features and bug fixes. This change will be effective the next time the services are deployed.

* Update @subql/cli version in package.json

The version of @subql/cli in the project's package.json file has been updated. The change involves an upgrade from version 4.0.5 to 4.13.1 to ensure compatibility with the latest features and improvements.

* Set non-unique index on authenticator field in schema.graphql

This was done for speed and it reflects a change the subquery team already made to the deployed graph

* Remove exodvs entities and handlers

* Deploy mainnet indexer

* Remove unique=false requirement

* Add publish script target

* Publish project